requirements.txt
astroid==2.5
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
decorator==4.4.2
entrypoints==0.3
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
flake8==3.7.7
Flask==1.1.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.1
lazy-object-proxy==1.5.2
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.5
Pillow==8.1.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.7.1
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tqdm==4.58.0
typed-ast==1.4.2
urllib3==1.26.3
validators==0.18.2
Werkzeug==1.0.1
wrapt==1.12.1
